943 add composite layers shortcuts
