+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuItem/MenuSwitch/SwitchItem/SwitchAction.tsx
@@ -1,0 +1,64 @@
+import {
+  Switch,
+  WithStyles,
+  createStyles,
+  withStyles,
+} from '@material-ui/core';
+import React, { ChangeEvent, memo, useCallback } from 'react';
+
+const styles = () =>
+  createStyles({
+    switch: {
+      marginRight: 2,
+    },
+    switchTrack: {
+      backgroundColor: '#E0E0E0',
+    },
+    switchBase: {
+      color: '#E0E0E0',
+      '&.Mui-checked': {
+        color: '#53888F',
+      },
+      '&.Mui-checked + .MuiSwitch-track': {
+        backgroundColor: '#B1D6DB',
+      },
+    },
+  });
+
+interface SwitchActionProps extends WithStyles<typeof styles> {
+  someLayerAreSelected: boolean;
+  toggleLayerValue: (selectedLayerId: string, checked: boolean) => void;
+  activeLayer: string;
+  validatedTitle: string;
+}
+const SwitchAction = ({
+  classes,
+  someLayerAreSelected,
+  toggleLayerValue,
+  activeLayer,
+  validatedTitle,
+}: SwitchActionProps) => {
+  const handleOnChangeSwitch = useCallback(
+    (event: ChangeEvent<HTMLInputElement>) => {
+      toggleLayerValue(activeLayer, event.target.checked);
+    },
+    [activeLayer, toggleLayerValue],
+  );
+  return (
+    <Switch
+      size="small"
+      className={classes.switch}
+      classes={{
+        switchBase: classes.switchBase,
+        track: classes.switchTrack,
+      }}
+      checked={someLayerAreSelected}
+      onChange={handleOnChangeSwitch}
+      inputProps={{
+        'aria-label': validatedTitle,
+      }}
+    />
+  );
+};
+
+export default memo(withStyles(styles)(SwitchAction));
