+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuItem/MenuSwitch/index.test.tsx
