+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuItem/index.tsx
@@ -1,4 +1,4 @@
-import React, { memo, useCallback, useEffect, useMemo, useState } from 'react';
+import React, { memo, useMemo } from 'react';
 import {
   createStyles,
   Grid,
@@ -7,16 +7,16 @@ import {
   AccordionDetails,
   AccordionSummary,
   makeStyles,
-  Chip,
 } from '@material-ui/core';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import { useSelector } from 'react-redux';
 import { LayersCategoryType } from 'config/types';
-import MenuSwitch from 'components/MapView/LeftPanel/layersPanel/MenuSwitch';
+import MenuSwitch from 'components/MapView/LeftPanel/layersPanel/MenuItem/MenuSwitch';
 import { useSafeTranslation } from 'i18n';
 import { Extent } from 'components/MapView/Layers/raster-utils';
 import { layersSelector } from 'context/mapStateSlice/selectors';
 import { filterActiveLayers } from 'components/MapView/utils';
+import SelectedLayersInformation from './SelectedLayersInformation';
 
 const useStyles = makeStyles(() =>
   createStyles({
@@ -34,9 +34,6 @@ const useStyles = makeStyles(() =>
     },
     summaryContent: {
       alignItems: 'center',
-    },
-    chipRoot: {
-      marginLeft: '3%',
     },
     title: {
       color: '#53888F',
@@ -57,74 +54,21 @@ const MenuItem = memo(({ title, layersCategories, extent }: MenuItemProps) => {
   const selectedLayers = useSelector(layersSelector);
   const classes = useStyles();
 
-  const categoryLayers = useMemo(() => {
-    return layersCategories
-      .map(layerCategory => {
-        return layerCategory.layers;
-      })
-      .flat();
-  }, [layersCategories]);
+  const categoryLayers = layersCategories
+    .map(layerCategory => {
+      return layerCategory.layers;
+    })
+    .flat();
 
-  const selectedCategoryLayers = useMemo(() => {
-    return selectedLayers.filter(layer => {
-      return categoryLayers.some(categoryLayer => {
-        return filterActiveLayers(layer, categoryLayer);
-      });
-    });
-  }, [categoryLayers, selectedLayers]);
-
-  const [informationChipLabel, setInformationChipLabel] = useState<string>(
-    selectedCategoryLayers.length.toString(),
+  const selectedCategoryLayers = useMemo(
+    () =>
+      selectedLayers.filter(layer =>
+        categoryLayers.some(categoryLayer =>
+          filterActiveLayers(layer, categoryLayer),
+        ),
+      ),
+    [categoryLayers, selectedLayers],
   );
-
-  useEffect(() => {
-    if (!selectedCategoryLayers.length) {
-      return;
-    }
-    setInformationChipLabel(selectedCategoryLayers.length.toString());
-  }, [selectedCategoryLayers.length]);
-
-  const renderedMenuSwitches = useMemo(() => {
-    return layersCategories.map((layerCategory: LayersCategoryType) => (
-      <MenuSwitch
-        key={layerCategory.title}
-        title={layerCategory.title}
-        layers={layerCategory.layers}
-        extent={extent}
-      />
-    ));
-  }, [extent, layersCategories]);
-
-  const handleChipOnMouseEnter = useCallback(() => {
-    setInformationChipLabel(
-      `${selectedCategoryLayers.length} ${t('Active Layer(s)')}`,
-    );
-  }, [selectedCategoryLayers.length, t]);
-
-  const handleChipOnMouseLeave = useCallback(() => {
-    setInformationChipLabel(selectedCategoryLayers.length.toString());
-  }, [selectedCategoryLayers.length]);
-
-  const renderedSelectedLayerInformation = useMemo(() => {
-    if (!selectedCategoryLayers.length) {
-      return null;
-    }
-    return (
-      <Chip
-        onMouseEnter={handleChipOnMouseEnter}
-        onMouseLeave={handleChipOnMouseLeave}
-        classes={{ root: classes.chipRoot }}
-        color="secondary"
-        label={informationChipLabel}
-      />
-    );
-  }, [
-    classes.chipRoot,
-    handleChipOnMouseEnter,
-    handleChipOnMouseLeave,
-    informationChipLabel,
-    selectedCategoryLayers.length,
-  ]);
 
   return (
     <Accordion elevation={0} classes={{ root: classes.root }}>
@@ -139,11 +83,20 @@ const MenuItem = memo(({ title, layersCategories, extent }: MenuItemProps) => {
         id={title}
       >
         <Typography classes={{ root: classes.title }}>{t(title)}</Typography>
-        {renderedSelectedLayerInformation}
+        <SelectedLayersInformation
+          selectedCategoryLayers={selectedCategoryLayers}
+        />
       </AccordionSummary>
       <AccordionDetails classes={{ root: classes.rootDetails }}>
         <Grid container direction="column">
-          {renderedMenuSwitches}
+          {layersCategories.map((layerCategory: LayersCategoryType) => (
+            <MenuSwitch
+              key={layerCategory.title}
+              title={layerCategory.title}
+              layers={layerCategory.layers}
+              extent={extent}
+            />
+          ))}
         </Grid>
       </AccordionDetails>
     </Accordion>
